✨ Feat: 상품 도메인 CRUD
- Exception 및 spring security 적용.

+++ src/main/kotlin/com/challengeteamkotlin/campdaddy/presentation/product/dto/response/ProductDetailResponse.kt
@@ -6,7 +6,7 @@ data class ProductDetailResponse(
     val title: String,
     val images: String,
     val content: String,
-    val pricePerDay: Long,
+    val pricePerDay: Int,
 ) {
     companion object {
         fun from(productEntity: ProductEntity) = ProductDetailResponse(
